+++ tuxemon/menu/theme.py
@@ -7,6 +7,7 @@ import pygame_menu
 from pygame_menu import locals, sound
 from pygame_menu.widgets.core.selection import Selection
 from pygame_menu.widgets.core.widget import Widget
+from pygame_menu.widgets.widget.menubar import MENUBAR_STYLE_ADAPTIVE
 
 from tuxemon import prepare
 from tuxemon.audio import get_sound_filename
@@ -107,6 +108,10 @@ def get_theme() -> pygame_menu.Theme:
         scrollbar_color=(235, 235, 235),
         scrollbar_slider_color=(200, 200, 200),
         widget_padding=(10, 20),
+        title_font_color=(10, 10, 10),
+        title_close_button=False,
+        title_bar_style=MENUBAR_STYLE_ADAPTIVE,
+        title_background_color=(197, 232, 234),
     )
 
     _theme = theme
