+++ tuxemon/states/choice/choice_npc.py
@@ -37,6 +37,7 @@ class ChoiceNpc(PygameMenuState):
         escape_key_exits: bool = False,
         **kwargs: Any,
     ) -> None:
+        width, height = prepare.SCREEN_SIZE
         theme = get_theme()
         if len(menu) > 12:
             theme.scrollarea_position = POSITION_EAST
@@ -66,7 +67,7 @@ class ChoiceNpc(PygameMenuState):
             self.menu.add.button(
                 name,
                 callback,
-                font_size=15,
+                font_size=round(0.012 * width),
                 align=ALIGN_CENTER,
                 selection_effect=HighlightSelection(),
             )
